+++ tower/CHANGELOG.md
@@ -10,6 +10,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
     response, and error types of the output service.
 - **builder**: Add `ServiceBuilder::check_service_clone` to check the output
     service can be cloned.
+- **buffer**: Add `Buffer::try_into_inner` for recovering inner service
 
 # 0.4.6 (February 26, 2021)